+++ oopz_sdk/services/__init__.py
@@ -77,14 +77,5 @@ class BaseService:
              timeout=timeout
         )
 
-    async def close(self) -> None:
-        await self.transport.close()
-
-    async def __aenter__(self):
-        return self
-
-    async def __aexit__(self, exc_type, exc, tb) -> None:
-        await self.close()
-
 
 __all__ = ["BaseService"]